requirements.txt
aiohttp>=3.9
beautifulsoup4>=4.12
dnspython>=2.4
ipwhois>=1.2
lxml>=4.9
paramiko>=3.3
python-nmap>=0.7
python-whois>=0.8
requests>=2.31
shodan>=1.31
sqlalchemy>=2.0
tabulate>=0.9
tqdm>=4.66
ujson>=5.8
xmltodict>=0.13
pyyaml>=6.0
python-dateutil>=2.8
